requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --no-dev --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
asgiref==3.10.0
    # via
    #   daphne
    #   django
    #   django-allauth
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.10.2
    # via daphne
automat==25.4.16
    # via twisted
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pyroscope-io
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   django-typer
    #   typer-slim
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   qrcode
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   fido2
    #   pyjwt
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via django-slices-starter
django==5.2.8
    # via
    #   django-allauth
    #   django-debug-toolbar
    #   django-ninja
    #   django-slices-starter
    #   django-typer
django-allauth==65.13.0
    # via django-slices-starter
django-debug-toolbar==6.1.0
    # via django-slices-starter
django-ninja==1.5.0
    # via django-slices-starter
django-typer==3.4.0
    # via django-slices-starter
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via pydantic
fido2==2.0.0
    # via django-allauth
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
graphql-core==3.2.7
    # via strawberry-graphql
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via django-slices-starter
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
importlib-metadata==8.7.0
    # via opentelemetry-api
incremental==24.7.2
    # via twisted
lia-web==0.2.3
    # via strawberry-graphql
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.3.1
    # via django-allauth
opentelemetry-api==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-psycopg
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   strawberry-graphql
opentelemetry-exporter-otlp==1.38.0
    # via django-slices-starter
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.38.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.38.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.59b0
    # via
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-psycopg
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-dbapi==0.59b0
    # via opentelemetry-instrumentation-psycopg
opentelemetry-instrumentation-django==0.59b0
    # via django-slices-starter
opentelemetry-instrumentation-httpx==0.59b0
    # via django-slices-starter
opentelemetry-instrumentation-psycopg==0.59b0
    # via django-slices-starter
opentelemetry-instrumentation-requests==0.59b0
    # via django-slices-starter
opentelemetry-instrumentation-wsgi==0.59b0
    # via opentelemetry-instrumentation-django
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   strawberry-graphql
opentelemetry-semantic-conventions==0.59b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.59b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   strawberry-graphql
protobuf==6.33.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psycopg==3.2.12
    # via django-slices-starter
psycopg-binary==3.2.12 ; implementation_name != 'pypy'
    # via psycopg
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.12.4
    # via
    #   django-ninja
    #   django-slices-starter
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via django-allauth
pyopenssl==25.3.0
    # via twisted
pyroscope-io==0.8.11
    # via django-slices-starter
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.2.1
    # via django-slices-starter
qrcode==8.2
    # via django-allauth
redis==7.0.1
    # via django-slices-starter
requests==2.32.5
    # via
    #   django-allauth
    #   opentelemetry-exporter-otlp-proto-http
restate-sdk==0.11.0
    # via django-slices-starter
rich==14.2.0
    # via django-typer
service-identity==24.2.0
    # via twisted
setuptools==80.9.0
    # via incremental
shellingham==1.5.4
    # via django-typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
strawberry-graphql==0.285.0
    # via django-slices-starter
twisted==25.5.0
    # via daphne
txaio==25.9.2
    # via autobahn
typer-slim==0.20.0
    # via django-typer
typing-extensions==4.15.0
    # via
    #   grpcio
    #   lia-web
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   strawberry-graphql
    #   twisted
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2 ; sys_platform == 'win32'
    # via
    #   django
    #   psycopg
urllib3==2.5.0
    # via requests
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-httpx
zipp==3.23.0
    # via importlib-metadata
zope-interface==8.1.1
    # via twisted
